Allow option middleware to overwrite entire options hash
