+++ lib/middleware.ts
@@ -7,6 +7,8 @@ const clientJs     = require('./client-js');
 const respMod      = require('resp-modifier');
 const mergeOpts    = require('./transform-options').mergeOptionsWithPlugins;
 var   count        = 0;
+const OPT_NAME     = 'middleware';
+
 
 export interface MiddlewareItem {
     id: string
@@ -40,7 +42,7 @@ mw.merge = function (options) {
  * @returns {Map}
  */
 mw.decorate = function (options) {
-    return options.update('middleware', x => x.map(createOne));
+    return options.update(OPT_NAME, x => x.map(createOne));
 };
 
 /**
@@ -64,11 +66,9 @@ module.exports.createOne = createOne;
 
 /**
  * JS -> Immutable transformation
- * @param coll
- * @returns {*}
  */
-mw.fromJS = function (coll) {
-    return Immutable.fromJS(coll.map(createOne));
+mw.fromJS = function (modified: MiddlewareItem[], options: Immutable.Map<string, any>) {
+    return options.set(OPT_NAME, Immutable.fromJS(modified.map(createOne)));
 };
 
 /**
